refactor: merge TestProject with WorkspaceProject
