+++ packages/vitest/src/node/pools/forks.ts
@@ -3,8 +3,8 @@ import type { RunnerRPC, RuntimeRPC } from '../../types/rpc'
 import type { ContextRPC, ContextTestEnvironment } from '../../types/worker'
 import type { Vitest } from '../core'
 import type { PoolProcessOptions, ProcessPool, RunWithFiles } from '../pool'
+import type { TestProject } from '../project'
 import type { SerializedConfig } from '../types/config'
-import type { WorkspaceProject } from '../workspace'
 import EventEmitter from 'node:events'
 import * as nodeos from 'node:os'
 import { resolve } from 'node:path'
@@ -16,7 +16,7 @@ import { wrapSerializableConfig } from '../../utils/config-helpers'
 import { envsOrder, groupFilesByEnv } from '../../utils/test-helpers'
 import { createMethodsRPC } from './rpc'
 
-function createChildProcessChannel(project: WorkspaceProject) {
+function createChildProcessChannel(project: TestProject) {
   const emitter = new EventEmitter()
   const cleanup = () => emitter.removeAllListeners()
 
@@ -99,7 +99,7 @@ export function createForksPool(
     let id = 0
 
     async function runFiles(
-      project: WorkspaceProject,
+      project: TestProject,
       config: SerializedConfig,
       files: string[],
       environment: ContextTestEnvironment,
@@ -153,8 +153,8 @@ export function createForksPool(
       // Cancel pending tasks from pool when possible
       ctx.onCancel(() => pool.cancelPendingTasks())
 
-      const configs = new WeakMap<WorkspaceProject, SerializedConfig>()
-      const getConfig = (project: WorkspaceProject): SerializedConfig => {
+      const configs = new WeakMap<TestProject, SerializedConfig>()
+      const getConfig = (project: TestProject): SerializedConfig => {
         if (configs.has(project)) {
           return configs.get(project)!
         }
