+++ packages/vitest/src/node/pools/vmThreads.ts
@@ -3,9 +3,9 @@ import type { RunnerRPC, RuntimeRPC } from '../../types/rpc'
 import type { ContextTestEnvironment } from '../../types/worker'
 import type { Vitest } from '../core'
 import type { PoolProcessOptions, ProcessPool, RunWithFiles } from '../pool'
+import type { TestProject } from '../project'
 import type { ResolvedConfig, SerializedConfig } from '../types/config'
 import type { WorkerContext } from '../types/worker'
-import type { WorkspaceProject } from '../workspace'
 import * as nodeos from 'node:os'
 import { resolve } from 'node:path'
 import { MessageChannel } from 'node:worker_threads'
@@ -18,7 +18,7 @@ import { createMethodsRPC } from './rpc'
 
 const suppressWarningsPath = resolve(rootDir, './suppress-warnings.cjs')
 
-function createWorkerChannel(project: WorkspaceProject) {
+function createWorkerChannel(project: TestProject) {
   const channel = new MessageChannel()
   const port = channel.port2
   const workerPort = channel.port1
@@ -98,7 +98,7 @@ export function createVmThreadsPool(
     let id = 0
 
     async function runFiles(
-      project: WorkspaceProject,
+      project: TestProject,
       config: SerializedConfig,
       files: string[],
       environment: ContextTestEnvironment,
@@ -156,8 +156,8 @@ export function createVmThreadsPool(
       // Cancel pending tasks from pool when possible
       ctx.onCancel(() => pool.cancelPendingTasks())
 
-      const configs = new Map<WorkspaceProject, SerializedConfig>()
-      const getConfig = (project: WorkspaceProject): SerializedConfig => {
+      const configs = new Map<TestProject, SerializedConfig>()
+      const getConfig = (project: TestProject): SerializedConfig => {
         if (configs.has(project)) {
           return configs.get(project)!
         }
