+++ packages/vitest/src/node/reporters/blob.ts
@@ -1,7 +1,7 @@
 import type { File } from '@vitest/runner'
 import type { Vitest } from '../core'
+import type { TestProject } from '../project'
 import type { Reporter } from '../types/reporter'
-import type { WorkspaceProject } from '../workspace'
 import { existsSync } from 'node:fs'
 import { mkdir, readdir, readFile, stat, writeFile } from 'node:fs/promises'
 import { parse, stringify } from 'flatted'
@@ -79,7 +79,7 @@ export class BlobReporter implements Reporter {
 export async function readBlobs(
   currentVersion: string,
   blobsDirectory: string,
-  projectsArray: WorkspaceProject[],
+  projectsArray: TestProject[],
 ) {
   // using process.cwd() because --merge-reports can only be used in CLI
   const resolvedDir = resolve(process.cwd(), blobsDirectory)
