+++ packages/vitest/src/node/reporters/github-actions.ts
@@ -1,7 +1,7 @@
 import type { File } from '@vitest/runner'
 import type { Vitest } from '../core'
+import type { TestProject } from '../project'
 import type { Reporter } from '../types/reporter'
-import type { WorkspaceProject } from '../workspace'
 import { stripVTControlCharacters } from 'node:util'
 import { getFullName, getTasks } from '@vitest/runner/utils'
 import { capturePrintError } from '../error'
@@ -16,7 +16,7 @@ export class GithubActionsReporter implements Reporter {
   onFinished(files: File[] = [], errors: unknown[] = []) {
     // collect all errors and associate them with projects
     const projectErrors = new Array<{
-      project: WorkspaceProject
+      project: TestProject
       title: string
       error: unknown
       file?: File
