+++ packages/vitest/src/node/spec.ts
@@ -1,13 +1,12 @@
 import type { SerializedTestSpecification } from '../runtime/types/utils'
-import type { TestProject } from './reported-workspace-project'
+import type { TestProject } from './project'
 import type { Pool } from './types/pool-options'
-import type { WorkspaceProject } from './workspace'
 
 export class TestSpecification {
   /**
    * @deprecated use `project` instead
    */
-  public readonly 0: WorkspaceProject
+  public readonly 0: TestProject
   /**
    * @deprecated use `moduleId` instead
    */
@@ -23,7 +22,7 @@ export class TestSpecification {
   // public readonly location: WorkspaceSpecLocation | undefined
 
   constructor(
-    workspaceProject: WorkspaceProject,
+    workspaceProject: TestProject,
     moduleId: string,
     pool: Pool,
     // location?: WorkspaceSpecLocation | undefined,
@@ -31,7 +30,7 @@ export class TestSpecification {
     this[0] = workspaceProject
     this[1] = moduleId
     this[2] = { pool }
-    this.project = workspaceProject.testProject
+    this.project = workspaceProject
     this.moduleId = moduleId
     this.pool = pool
     // this.location = location
@@ -53,7 +52,7 @@ export class TestSpecification {
    * @deprecated
    */
   *[Symbol.iterator]() {
-    yield this.project.workspaceProject
+    yield this.project
     yield this.moduleId
     yield this.pool
   }
