+++ packages/vitest/src/node/state.ts
@@ -1,6 +1,6 @@
 import type { File, Task, TaskResultPack } from '@vitest/runner'
 import type { UserConsoleLog } from '../types/general'
-import type { WorkspaceProject } from './workspace'
+import type { TestProject } from './project'
 import { createFileTask } from '@vitest/runner/utils'
 import { TestCase, TestModule, TestSuite } from './reporters/reported-tasks'
 
@@ -106,7 +106,7 @@ export class StateManager {
     })
   }
 
-  collectFiles(project: WorkspaceProject, files: File[] = []) {
+  collectFiles(project: TestProject, files: File[] = []) {
     files.forEach((file) => {
       const existing = this.filesMap.get(file.filepath) || []
       const otherFiles = existing.filter(
@@ -127,7 +127,7 @@ export class StateManager {
   }
 
   clearFiles(
-    project: WorkspaceProject,
+    project: TestProject,
     paths: string[] = [],
   ) {
     paths.forEach((path) => {
@@ -157,7 +157,7 @@ export class StateManager {
     })
   }
 
-  updateId(task: Task, project: WorkspaceProject) {
+  updateId(task: Task, project: TestProject) {
     if (this.idMap.get(task.id) === task) {
       return
     }
@@ -214,7 +214,7 @@ export class StateManager {
     ).length
   }
 
-  cancelFiles(files: string[], project: WorkspaceProject) {
+  cancelFiles(files: string[], project: TestProject) {
     this.collectFiles(
       project,
       files.map(filepath =>
