+++ packages/vitest/src/typecheck/collect.ts
@@ -1,6 +1,6 @@
 import type { File, Suite, Test } from '@vitest/runner'
 import type { RawSourceMap } from 'vite-node'
-import type { WorkspaceProject } from '../node/workspace'
+import type { TestProject } from '../node/project'
 import {
   calculateSuiteHash,
   generateHash,
@@ -44,7 +44,7 @@ export interface FileInformation {
 }
 
 export async function collectTests(
-  ctx: WorkspaceProject,
+  ctx: TestProject,
   filepath: string,
 ): Promise<null | FileInformation> {
   const request = await ctx.vitenode.transformRequest(filepath, filepath)
